update: Resolve Clippy errors (#87)
This commit addresses several Clippy errors reported by cargo clippy.
The majority of the errors have been resolved, leaving only two remaining issues: one related to the unimplemented USB Transceiver buffer and another concerning the introduction of a new function in InactiveFilter.

+++ crates/crabe_io/src/league/real/task.rs
@@ -25,7 +25,7 @@ impl Real {
         let mut packet = PcToBase::default();
         for (id, command) in commands {
             let (kicker_cmd, kick_power) = match command.kick {
-                None => (Kicker::NoKick, 0.0 as f32),
+                None => (Kicker::NoKick, 0.0_f32),
                 Some(Kick::StraightKick { power }) => (Kicker::Flat, power),
                 Some(Kick::ChipKick { power }) => (Kicker::Chip, power),
             };
